Added code for issue 28
